Started logic division
