+++ Makefile.am
@@ -1,5 +1,5 @@
 bin_PROGRAMS = openstratos
-openstratos_SOURCES = openstratos.cc utils.cc threads.cc camera/Camera.cc gps/GPS.cc serial/Serial.cc logger/Logger.cc gsm/GSM.cc
+openstratos_SOURCES = openstratos.cc logic/main_while.cc logic/shut_down.cc utils.cc threads.cc camera/Camera.cc gps/GPS.cc serial/Serial.cc logger/Logger.cc gsm/GSM.cc
 openstratos_CPPFLAGS = -std=c++14
 
 EXTRA_PROGRAMS = utesting
